+++ datafusion/physical-expr/Cargo.toml
@@ -77,7 +77,3 @@ tokio = { workspace = true, features = ["rt-multi-thread"] }
 [[bench]]
 harness = false
 name = "in_list"
-
-[[bench]]
-harness = false
-name = "concat"
