+++ datafusion/physical-expr/src/lib.rs
@@ -28,7 +28,6 @@ mod partitioning;
 mod physical_expr;
 pub mod planner;
 mod scalar_function;
-pub mod string_expressions;
 pub mod udf;
 pub mod utils;
 pub mod window;
