+++ datafusion/proto/src/logical_plan/from_proto.rs
@@ -37,9 +37,8 @@ use datafusion_expr::expr::Unnest;
 use datafusion_expr::expr::{Alias, Placeholder};
 use datafusion_expr::window_frame::{check_window_frame, regularize_window_order_by};
 use datafusion_expr::{
-    coalesce, concat_expr, concat_ws_expr, ends_with,
+    coalesce,
     expr::{self, InList, Sort, WindowFunction},
-    initcap,
     logical_plan::{PlanType, StringifiedPlan},
     AggregateFunction, Between, BinaryExpr, BuiltInWindowFunction, BuiltinScalarFunction,
     Case, Cast, Expr, GetFieldAccess, GetIndexedField, GroupingSet,
@@ -418,10 +417,6 @@ impl From<&protobuf::ScalarFunction> for BuiltinScalarFunction {
         use protobuf::ScalarFunction;
         match f {
             ScalarFunction::Unknown => todo!(),
-            ScalarFunction::Concat => Self::Concat,
-            ScalarFunction::ConcatWithSeparator => Self::ConcatWithSeparator,
-            ScalarFunction::EndsWith => Self::EndsWith,
-            ScalarFunction::InitCap => Self::InitCap,
             ScalarFunction::Coalesce => Self::Coalesce,
         }
     }
@@ -1287,19 +1282,6 @@ pub fn parse_expr(
 
             match scalar_function {
                 ScalarFunction::Unknown => Err(proto_error("Unknown scalar function")),
-                ScalarFunction::InitCap => {
-                    Ok(initcap(parse_expr(&args[0], registry, codec)?))
-                }
-                ScalarFunction::Concat => {
-                    Ok(concat_expr(parse_exprs(args, registry, codec)?))
-                }
-                ScalarFunction::ConcatWithSeparator => {
-                    Ok(concat_ws_expr(parse_exprs(args, registry, codec)?))
-                }
-                ScalarFunction::EndsWith => Ok(ends_with(
-                    parse_expr(&args[0], registry, codec)?,
-                    parse_expr(&args[1], registry, codec)?,
-                )),
                 ScalarFunction::Coalesce => {
                     Ok(coalesce(parse_exprs(args, registry, codec)?))
                 }
