+++ datafusion/proto/src/logical_plan/to_proto.rs
@@ -1407,10 +1407,6 @@ impl TryFrom<&BuiltinScalarFunction> for protobuf::ScalarFunction {
 
     fn try_from(scalar: &BuiltinScalarFunction) -> Result<Self, Self::Error> {
         let scalar_function = match scalar {
-            BuiltinScalarFunction::Concat => Self::Concat,
-            BuiltinScalarFunction::ConcatWithSeparator => Self::ConcatWithSeparator,
-            BuiltinScalarFunction::EndsWith => Self::EndsWith,
-            BuiltinScalarFunction::InitCap => Self::InitCap,
             BuiltinScalarFunction::Coalesce => Self::Coalesce,
         };
